requirements.txt
streamlit
pandas
numpy
joblib
fpdf
azure-storage-blob